Exhaustive state conditions for all realtime operations.
When calling `Connection.close()` before reaching the CONNECTED state,
the iOS library was trying to send a WebSocket message before the
handshake was even finished. That caused it to crash sometimes.

That situation wasn't contemplated on the spec. I've gone through it
filling a table and filling the gaps, so that the spec is exhaustive
about all possible combinations of state and operations.

(So, yes, https://github.com/zwopple/PocketSocket/pull/48 wasn't fixing
the right problem.)