Copy stats types information from Ruby documentation
My initial motivation here is EDX-175. We want the entire public API of
our SDKs to have docstring comments. We will be approaching this by
using this spec’s IDL as our source of truth for what constitutes our
SDK’s public API, to which tech writers will then add generic
descriptive comments. Hence, the IDL will need to contain the
stats-related types, which it currently doesn’t. The first step towards
achieving this is to make sure that all of these types are documented in
the spec.

Currently, the types are not documented, instead pointing the reader
towards the Ruby documentation. I don’t think it makes sense for a
generic feature spec to refer to a specific implementation as a
reference - seems a bit circular, and requires the reader to have at
least a passing knowledge of that language. Furthermore, it means that
there’s a chunk of the feature spec that won’t go through the usual
feature spec review process - any changes to that part of the Ruby SDK
will become changes to the feature spec, without the Ruby SDK developers
necessarily even realising this. Also, at the time of writing, the Ruby
SDK generated documentation was broken, not displaying any classes at
all (all of the links currently in the feature spec giving 404s).

The lack of stats documentation has also caused us to be blocked on
https://github.com/ably/ably-flutter/issues/106 (adding stats
functionality to the Flutter SDK), because of inconsistencies between
ably-cocoa and ably-java which the current documentation was not
sufficient to help us resolve.

The information and comments that I’ve added here are taken from
ably-ruby@f5eac15. I’ve lightly edited the descriptions, but I think
they still could be improved a lot. I think I’ll need input from someone
with more knowledge of what the numbers in the stats response actually
mean, though.

There are a couple of types mentioned in the spec which are not actually
implemented in the Ruby SDK – PushStats and XchgMessages. I’ll handle
those separately.